+++ src/ts/Ads/Managers/AdRequestManager.ts
@@ -27,6 +27,12 @@ import { CampaignParser } from 'Ads/Parsers/CampaignParser';
 import { SDKMetrics, LoadV5 } from 'Ads/Utilities/SDKMetrics';
 import { RequestError } from 'Core/Errors/RequestError';
 import { SdkStats } from 'Ads/Utilities/SdkStats';
+import { Observable2 } from 'Core/Utilities/Observable';
+
+export interface INotCachedLoadedCampaign {
+    notCachedCampaign: Campaign;
+    notCachedTrackingUrls: ICampaignTrackingUrls;
+}
 
 export interface IParsedPlacementPreloadData {
     campaignAvailable: boolean;
@@ -89,6 +95,8 @@ export class AdRequestManager extends CampaignManager {
     private _deviceFreeSpace: number;
     private _userPrivacyManager: UserPrivacyManager;
     private _currentSession: Session | null;
+
+    public readonly onAdditionalPlacementsReady = new Observable2<string | undefined, IPlacementIdMap<INotCachedLoadedCampaign | undefined>>();
 
     constructor(platform: Platform, core: ICore, coreConfig: CoreConfiguration, adsConfig: AdsConfiguration, assetManager: AssetManager, sessionManager: SessionManager, adMobSignalFactory: AdMobSignalFactory, request: RequestManager, clientInfo: ClientInfo, deviceInfo: DeviceInfo, metaDataManager: MetaDataManager, cacheBookkeeping: CacheBookkeepingManager, contentTypeHandlerManager: ContentTypeHandlerManager, privacySDK: PrivacySDK, userPrivacyManager: UserPrivacyManager) {
         super();
@@ -168,7 +176,7 @@ export class AdRequestManager extends CampaignManager {
             retryDelay: 0,
             followRedirects: false,
             retryWithConnectionEvents: false,
-            timeout: 10000
+            timeout: 20000
         })).then((response) => {
             if (response) {
                 SdkStats.increaseAdRequestOrdinal();
@@ -186,7 +194,7 @@ export class AdRequestManager extends CampaignManager {
         });
     }
 
-    public requestLoad(placementId: string, rescheduled: boolean = false): Promise<ILoadedCampaign | undefined> {
+    public requestLoad(placementId: string, additionalPlacements: string[] = [], rescheduled: boolean = false): Promise<ILoadedCampaign | undefined> {
         // setting that placementId is being loaded,
         // this is used to track load cancellation due to reload request
         this._ongoingLoadRequests[placementId] = true;
@@ -195,7 +203,7 @@ export class AdRequestManager extends CampaignManager {
         if (this._ongoingPreloadRequest !== null) {
             let promiseResolve: () => void;
             const promise = new Promise((resolve) => { promiseResolve = resolve; }).then(() =>
-                this.requestLoad(placementId, true)
+                this.requestLoad(placementId, additionalPlacements, true)
             );
 
             this._ongoingPreloadRequest = this._ongoingPreloadRequest.then(() => { promiseResolve(); });
@@ -207,7 +215,7 @@ export class AdRequestManager extends CampaignManager {
         if (this._ongoingReloadRequest !== null) {
             let promiseResolve: () => void;
             const promise = new Promise((resolve) => { promiseResolve = resolve; }).then(() =>
-                this.requestLoad(placementId, true)
+                this.requestLoad(placementId, additionalPlacements, true)
             );
 
             this._ongoingReloadRequest = this._ongoingReloadRequest.then(() => { promiseResolve(); });
@@ -253,7 +261,7 @@ export class AdRequestManager extends CampaignManager {
                 CampaignManager.createRequestUrl(this.getBaseUrl(), this._platform, this._clientInfo, this._deviceInfo, this._coreConfig, this._lastAuctionId, false),
                 CampaignManager.createRequestBody(this._clientInfo, this._coreConfig, this._deviceInfo, this._userPrivacyManager, this._sessionManager, this._privacy, undefined, fullyCachedCampaignIds, versionCode, this._adMobSignalFactory, freeSpace, this._metaDataManager, this._adsConfig, true, this.getPreviousPlacementId(), requestPrivacy, legacyRequestPrivacy, false, this._adsConfig.getPlacement(placementId))
             ]);
-        }).then(([requestUrl, requestBody]) => this._request.post(requestUrl, JSON.stringify(this.makeLoadBody(<ILoadV5BodyExtra>requestBody, placementId)), [], {
+        }).then(([requestUrl, requestBody]) => this._request.post(requestUrl, JSON.stringify(this.makeLoadBody(<ILoadV5BodyExtra>requestBody, placementId, additionalPlacements)), [], {
             retries: 0,
             retryDelay: 0,
             followRedirects: false,
@@ -269,7 +277,7 @@ export class AdRequestManager extends CampaignManager {
                 return this.requestLoad(placementId);
             }
             SDKMetrics.reportMetricEvent(LoadV5.LoadRequestParsingResponse);
-            return this.parseLoadResponse(response, this._adsConfig.getPlacement(placementId));
+            return this.parseLoadResponse(response, this._adsConfig.getPlacement(placementId), additionalPlacements);
         }).then((campaign) => {
             delete this._ongoingLoadRequests[placementId];
             if (campaign) {
@@ -328,7 +336,7 @@ export class AdRequestManager extends CampaignManager {
             retryDelay: 0,
             followRedirects: false,
             retryWithConnectionEvents: false,
-            timeout: 10000
+            timeout: 20000
         })).then((response) => {
             if (response) {
                 SdkStats.increaseAdRequestOrdinal();
@@ -352,8 +360,19 @@ export class AdRequestManager extends CampaignManager {
         return this.requestPreload();
     }
 
+    public loadCampaignWithAdditionalPlacement(placement: Placement): Promise<ILoadedCampaign | undefined> {
+        let additionalPlacements: string[] = [];
+
+        if (placement.hasAdUnitId()) {
+            additionalPlacements = this._adsConfig.getPlacementsForAdunit(placement.getAdUnitId()!)
+                .filter(placementId => placementId !== placement.getId());
+        }
+
+        return this.requestLoad(placement.getId(), additionalPlacements);
+    }
+
     public loadCampaign(placement: Placement): Promise<ILoadedCampaign | undefined> {
-        return this.requestLoad(placement.getId());
+        return this.requestLoad(placement.getId(), []);
     }
 
     public isPreloadDataExpired(): boolean {
@@ -398,7 +417,7 @@ export class AdRequestManager extends CampaignManager {
         return Promise.resolve();
     }
 
-    private parseLoadResponse(response: INativeResponse, placement: Placement): Promise<ILoadedCampaign | undefined> {
+    private parseLoadResponse(response: INativeResponse, placement: Placement, additionalPlacements: string[]): Promise<ILoadedCampaign | undefined> {
         // time
         let json: IRawAuctionV5Response;
         try {
@@ -418,7 +437,24 @@ export class AdRequestManager extends CampaignManager {
             return Promise.reject(new AdRequestManagerError('No placement', 'no_plc'));
         }
 
-        return this.createLoadedCampaign(json, placement, auctionStatusCode);
+        return Promise.all(
+            // Skip caching for those campaigns since we don't need them immediately
+            additionalPlacements.map((x) => this.createNotCachedLoadedCampaign(json, this._adsConfig.getPlacement(x), auctionStatusCode).catch((err) => {
+                return undefined;
+            }))
+        ).then((loadedCampaigns) => {
+            const additionalCampaigns = loadedCampaigns.reduce<IPlacementIdMap<INotCachedLoadedCampaign | undefined>>((previousValue, currentValue, currentIndex) => {
+                previousValue[additionalPlacements[currentIndex]] = currentValue;
+                return previousValue;
+            }, {});
+            this.onAdditionalPlacementsReady.trigger(placement.getAdUnitId(), additionalCampaigns);
+        }).catch(() => {
+            // Skip any errors so that we can handle actual load request
+        }).then(
+            () => this.createNotCachedLoadedCampaign(json, placement, auctionStatusCode)
+        ).then(
+            (notCachedLoadedCampaign) => this.cacheCampaign(notCachedLoadedCampaign)
+        );
     }
 
     private parsePreloadData(response: IRawAuctionV5Response): IPlacementIdMap<IParsedPlacementPreloadData> | null {
@@ -476,11 +512,6 @@ export class AdRequestManager extends CampaignManager {
             } else {
                 throw new AdRequestManagerError('Failed to read campaign', 'no_campaign');
             }
-        }).then(campaign => {
-            return this._assetManager.setup(campaign).catch((err) => {
-                // If caching failed, we still can stream an ad.
-                return campaign;
-            });
         });
     }
 
@@ -502,7 +533,7 @@ export class AdRequestManager extends CampaignManager {
         return Promise.resolve(trackingUrls);
     }
 
-    private createLoadedCampaign(response: IRawAuctionV5Response, placement: Placement, auctionStatusCode: AuctionStatusCode): Promise<ILoadedCampaign | undefined> {
+    private createNotCachedLoadedCampaign(response: IRawAuctionV5Response, placement: Placement, auctionStatusCode: AuctionStatusCode): Promise<INotCachedLoadedCampaign | undefined> {
         return this.parseMediaAndTrackingUrls(response, placement, auctionStatusCode).then(({ mediaId, trackingId }) => {
             return Promise.all([
                 this.parseTrackingUrls(response, trackingId, auctionStatusCode),
@@ -512,7 +543,7 @@ export class AdRequestManager extends CampaignManager {
             if (!campaign || !trackingUrls) {
                 return Promise.resolve(undefined);
             }
-            return Promise.resolve({ campaign: campaign, trackingUrls: trackingUrls });
+            return Promise.resolve({ notCachedCampaign: campaign, notCachedTrackingUrls: trackingUrls });
         });
     }
 
@@ -540,6 +571,22 @@ export class AdRequestManager extends CampaignManager {
         return Promise.resolve({
             mediaId,
             trackingId
+        });
+    }
+
+    public cacheCampaign(notCachedLoadedCampaign: INotCachedLoadedCampaign | undefined): Promise<ILoadedCampaign | undefined> {
+        if (notCachedLoadedCampaign === undefined) {
+            return Promise.resolve(undefined);
+        }
+
+        return this._assetManager.setup(notCachedLoadedCampaign.notCachedCampaign).catch((err) => {
+            // If caching failed, we still can stream an ad.
+            return notCachedLoadedCampaign.notCachedCampaign;
+        }).then((campaign) => {
+            return {
+                campaign: campaign,
+                trackingUrls: notCachedLoadedCampaign.notCachedTrackingUrls
+            };
         });
     }
 
@@ -573,10 +620,14 @@ export class AdRequestManager extends CampaignManager {
         }
 
         return Promise.all(
-            placementsToLoad.map((x) => this.createLoadedCampaign(json, x, auctionStatusCode).catch((err) => {
-                this.handleError(LoadV5.ReloadRequestParseCampaignFailed, err);
-                return undefined;
-            }))
+            placementsToLoad.map(
+                (x) => this.createNotCachedLoadedCampaign(json, x, auctionStatusCode)
+                    .then((loadedCampaign) => this.cacheCampaign(loadedCampaign)).catch((err) => {
+                        this.handleError(LoadV5.ReloadRequestParseCampaignFailed, err);
+                        return undefined;
+                    }
+                )
+            )
         ).then((loadedCampaigns) => {
             loadedCampaigns.forEach((loadedCampaign, index) => {
                 const placementId = placementsToLoad[index].getId();
@@ -626,15 +677,37 @@ export class AdRequestManager extends CampaignManager {
         return body;
     }
 
-    private makeLoadBody(body: ILoadV5BodyExtra, placementId: string): unknown {
+    private makeLoadBody(body: ILoadV5BodyExtra, placementId: string, additionalPlacements: string[]): unknown {
         const preloadData: IPlacementIdMap<IParsedPlacementPreloadData> = {};
-        if (this._preloadData !== null && this._preloadData[placementId] !== undefined) {
-            preloadData[placementId] = this._preloadData[placementId];
+
+        if (this._preloadData !== null) {
+            const tempPreloadData = this._preloadData;
+
+            if (tempPreloadData !== undefined) {
+                preloadData[placementId] = tempPreloadData[placementId];
+            }
+
+            additionalPlacements.reduce((previousValue, currentValue) => {
+                if (tempPreloadData[currentValue] !== undefined) {
+                    previousValue[currentValue] = tempPreloadData[currentValue];
+                }
+                return previousValue;
+            }, preloadData);
         }
 
         if (this._currentSession === null) {
             throw new AdRequestManagerError('Current session is not set', 'no_session');
         }
+
+        additionalPlacements.reduce((previousValue, currentValue) => {
+            const placement = this._adsConfig.getPlacement(currentValue);
+            previousValue[currentValue] = {
+                adTypes: placement.getAdTypes(),
+                allowSkip: placement.allowSkip(),
+                auctionType: placement.getAuctionType()
+            };
+            return previousValue;
+        }, body.placements);
 
         body.auctionId = this._currentSession.getId();
         body.load = true;