+++ src/ts/Ads/Managers/MediationLoadTrackingManager.ts
@@ -6,7 +6,7 @@ import { GameSessionCounters } from 'Ads/Utilities/GameSessionCounters';
 const INITIAL_AD_REQUEST_WAIT_TIME_IN_MS = 250;
 
 export enum MediationExperimentType {
-    CacheModeAllowed = 'cma2',
+    DynamicCacheMode = 'dcm',
     AuctionXHR = 'xhr',
     LoadV5 = 'lv5',
     None = 'none'